requirements.txt
contourpy==1.0.6
cvxopt==1.3.0
cycler==0.11.0
fonttools==4.38.0
Fred-Frechet==1.9.25
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.24.1
packaging==23.0
pandas==1.5.2
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
scikit-learn==1.2.0
scipy==1.10.0
six==1.16.0
threadpoolctl==3.1.0
